+++ notifications/templatetags/notifications_tags.py
@@ -4,6 +4,7 @@ from django import get_version
 from django.core.cache import cache
 from django.template import Library
 from django.utils.html import format_html
+from django.utils.safestring import mark_safe
 from packaging.version import (
     parse as parse_version,  # pylint: disable=no-name-in-module,import-error
 )
@@ -95,7 +96,7 @@ def register_notify_callbacks(badge_class='live_notify_badge',  # pylint: disabl
     for callback in callbacks.split(','):
         script += "register_notifier(" + callback + ");"
     script += "</script>"
-    return format_html(script)
+    return mark_safe(script)
 
 
 @register.simple_tag(takes_context=True)
@@ -104,16 +105,16 @@ def live_notify_badge(context, badge_class='live_notify_badge'):
     if not user:
         return ''
 
-    html = "<span class='{badge_class}'>{unread}</span>".format(
-        badge_class=badge_class, unread=get_cached_notification_unread_count(user)
+    return format_html(
+        "<span class='{}'>{}</span>",
+        badge_class,
+        get_cached_notification_unread_count(user),
     )
-    return format_html(html)
 
 
 @register.simple_tag
 def live_notify_list(list_class='live_notify_list'):
-    html = "<ul class='{list_class}'></ul>".format(list_class=list_class)
-    return format_html(html)
+    return format_html("<ul class='{}'></ul>", list_class)
 
 
 def user_context(context):